Add initial documentation setup and license file
This commit introduces a new GitHub Actions workflow for deploying
documentation to GitHub Pages. The workflow triggers on push events
to the main branch for specific paths related to documentation.
Additionally, a .envrc file is added for environment variable
management, and a LICENSE.txt file is introduced to clarify
the project's licensing under the MIT License. The .gitignore file
is also updated to exclude the 'site' directory generated by
MkDocs.